Flatdata's Structs wrap a single byte now.
A flatdata struct consists of the first byte to the data. For
reading and write, we take the address of this byte. This approach
allows us to define ArrayView as an interface over slice. Also,
there is no need to distinguish between mutable and non-mutable
structs, since they are directly represented as &T and &mut T.