Revamp flatdata's types design from scratch.
Struct's come now in two flavours const and mut. Both are not bound
with lifetimes. Instead, all containers return Handle<'a, T> or
HandleMut<'a, T>, where T is a flatdata type, dependending on
whether the data will be read or writting. Handle{Mut} implement
AsRef, Deref and DerefMut to allow transparent usage of the
underlying functions.